+++ packages/contracts-communication/contracts/apps/examples/ExampleAppV1.sol
@@ -10,7 +10,7 @@ import {Address} from "@openzeppelin/contracts/utils/Address.sol";
 
 contract ExampleAppV1 is ICAppV1 {
     event MessageReceived(uint64 srcChainId, bytes32 sender, uint256 dbNonce, uint64 entryIndex, bytes message);
-    event MessageSent(uint256 dstChainId, uint256 dbNonce, uint64 entryIndex, bytes32 transactionId);
+    event MessageSent(uint64 dstChainId, uint256 dbNonce, uint64 entryIndex, bytes32 transactionId);
 
     constructor(address admin) ICAppV1(admin) {
         _grantRole(IC_GOVERNOR_ROLE, admin);
@@ -23,7 +23,7 @@ contract ExampleAppV1 is ICAppV1 {
 
     /// @notice Sends a basic message to the destination chain.
     function sendMessage(
-        uint256 dstChainId,
+        uint64 dstChainId,
         uint256 gasLimit,
         uint256 gasAirdrop,
         bytes calldata message
@@ -42,7 +42,7 @@ contract ExampleAppV1 is ICAppV1 {
 
     /// @notice Returns the fee required to send a message using `sendMessage`.
     function getMessageFee(
-        uint256 dstChainId,
+        uint64 dstChainId,
         uint256 gasLimit,
         uint256 gasAirdrop,
         bytes memory message
