+++ packages/contracts-communication/contracts/interfaces/IInterchainModule.sol
@@ -18,10 +18,10 @@ interface IInterchainModule {
     /// @dev Could be only called by the Interchain DataBase contract.
     /// @param dstChainId       The chain id of the destination chain
     /// @param versionedBatch   The versioned batch to verify
-    function requestBatchVerification(uint256 dstChainId, bytes memory versionedBatch) external payable;
+    function requestBatchVerification(uint64 dstChainId, bytes memory versionedBatch) external payable;
 
     /// @notice Get the Module fee for verifying a batch on the specified destination chain.
     /// @param dstChainId   The chain id of the destination chain
     /// @param dbNonce      The database nonce of the batch on the source chain
-    function getModuleFee(uint256 dstChainId, uint256 dbNonce) external view returns (uint256);
+    function getModuleFee(uint64 dstChainId, uint256 dbNonce) external view returns (uint256);
 }
