+++ packages/contracts-communication/test/InterchainClientV1.Src.t.sol
@@ -67,14 +67,14 @@ contract InterchainClientV1SourceTest is InterchainClientV1BaseTest {
     }
 
     /// @dev Override the DB's returned interchain fee for the given destination chain and modules.
-    function mockInterchainFee(uint256 dstChainId, address[] memory modules, uint256 interchainFee) internal {
+    function mockInterchainFee(uint64 dstChainId, address[] memory modules, uint256 interchainFee) internal {
         vm.mockCall(
             icDB, abi.encodeCall(InterchainDBMock.getInterchainFee, (dstChainId, modules)), abi.encode(interchainFee)
         );
     }
 
     /// @dev Override the ExecutionService's returned execution fee for the given destination chain and transaction.
-    function mockExecutionFee(uint256 dstChainId, InterchainTransaction memory icTx, uint256 executionFee) internal {
+    function mockExecutionFee(uint64 dstChainId, InterchainTransaction memory icTx, uint256 executionFee) internal {
         uint256 txPayloadSize = getEncodedTx(icTx).length;
         vm.mockCall(
             execService,
