+++ packages/contracts-communication/test/harnesses/InterchainTransactionLibHarness.sol
@@ -6,7 +6,7 @@ import {InterchainTransaction, InterchainTransactionLib} from "../../contracts/l
 contract InterchainTransactionLibHarness {
     function constructLocalTransaction(
         address srcSender,
-        uint256 dstChainId,
+        uint64 dstChainId,
         bytes32 dstReceiver,
         uint256 dbNonce,
         uint64 entryIndex,
